Allow Modbus to be configured using either "old" config or discovery_info.
Simplify __init__.py to not contain code, but only the schemas.
