+++ homeassistant/components/modbus/const.py
@@ -16,7 +16,7 @@ CONF_PRECISION = "precision"
 CONF_COILS = "coils"
 
 # integration names
-DEFAULT_HUB = "default"
+DEFAULT_HUB = "modbusHub"
 MODBUS_DOMAIN = "modbus"
 
 # data types
@@ -67,6 +67,7 @@ CONF_VERIFY_STATE = "verify_state"
 
 # climate.py
 CONF_CLIMATES = "climates"
+CONF_CLIMATE = "climate"
 CONF_TARGET_TEMP = "target_temp_register"
 CONF_CURRENT_TEMP = "current_temp_register"
 CONF_CURRENT_TEMP_REGISTER_TYPE = "current_temp_register_type"
@@ -80,6 +81,7 @@ DEFAULT_STRUCTURE_PREFIX = ">f"
 DEFAULT_TEMP_UNIT = "C"
 
 # cover.py
+CONF_COVER = "cover"
 CONF_STATE_OPEN = "state_open"
 CONF_STATE_CLOSED = "state_closed"
 CONF_STATE_OPENING = "state_opening"
